fix some bugs and clean up
