+++ acl/delegation.acl
@@ -4,6 +4,7 @@
 #[bypass_all]
 
 [bypass_list]
+127.0.0.1
 (^|\.)bing\.com$
 
 [proxy_list: Seattle, WA]
