--- src/local.h
@@ -1,92 +1,0 @@
-/*
- * local.h - Define the client's buffers and callbacks
- *
- * Copyright (C) 2013 - 2019, Max Lv <max.c.lv@gmail.com>
- *
- * This file is part of the shadowsocks-libev.
- *
- * shadowsocks-libev is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3 of the License, or
- * (at your option) any later version.
- *
- * shadowsocks-libev is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with shadowsocks-libev; see the file COPYING. If not, see
- * <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _LOCAL_H
-#define _LOCAL_H
-
-#include <libcork/ds.h>
-
-#ifdef HAVE_LIBEV_EV_H
-#include <libev/ev.h>
-#else
-#include <ev.h>
-#endif
-
-#ifdef __MINGW32__
-#include "winsock.h"
-#endif
-
-#include "crypto.h"
-#include "jconf.h"
-#include "protocol.h"
-
-#include "common.h"
-
-typedef struct server_ctx {
-    ev_io io;
-    int connected;
-    struct server *server;
-} server_ctx_t;
-
-typedef struct server {
-    int fd;
-    int stage;
-
-    struct server_ctx *recv_ctx;
-    struct server_ctx *send_ctx;
-    struct listen_ctx *listen_ctx;
-    struct remote *remote;
-
-    buffer_t *buf;
-    buffer_t *abuf;
-
-    struct cork_dllist_item entries;
-} server_t;
-
-typedef struct remote_ctx {
-    ev_io io;
-    ev_timer watcher;
-    int connected;
-    struct remote *remote;
-} remote_ctx_t;
-
-typedef struct remote {
-    int fd;
-    int direct;
-#ifdef TCP_FASTOPEN_WINSOCK
-    OVERLAPPED olap;
-    int connect_ex_done;
-#endif
-
-    buffer_t *buf;
-
-    crypto_t *crypto;
-    cipher_ctx_t *e_ctx;
-    cipher_ctx_t *d_ctx;
-
-    struct remote_ctx *recv_ctx;
-    struct remote_ctx *send_ctx;
-    struct server *server;
-    struct sockaddr_storage *addr;
-} remote_t;
-
-#endif // _LOCAL_H
