+++ src/server.c
@@ -77,9 +77,6 @@
 
 #endif
 
-static void signal_cb(EV_P_ ev_signal *w, int revents);
-static void accept_cb(EV_P_ ev_io *w, int revents);
-
 static remote_t *
 connect_to_remote(EV_P_ server_t *server, struct sockaddr_storage *addr);
 
@@ -92,128 +89,7 @@ int no_delay = 0;
 int fast_open = 0;
 int ipv6first = 0;
 
-static int ret_val   = 0;
-
-static char *manager_addr = NULL;
 uint64_t tx = 0, rx = 0;
-
-#ifndef __MINGW32__
-ev_timer stat_watcher;
-#endif
-
-struct ev_signal sigint_watcher;
-struct ev_signal sigterm_watcher;
-#ifndef __MINGW32__
-struct ev_signal sigchld_watcher;
-#endif
-
-static struct cork_dllist listeners;
-
-#ifndef __MINGW32__
-static void
-stat_update_cb(EV_P_ ev_timer *watcher, int revents)
-{
-    listen_ctx_t *listen_ctx
-        = cork_container_of(watcher, listen_ctx_t, stat_watcher);
-    struct sockaddr_storage addr;
-
-    if (getsockname(listen_ctx->fd,
-                    (struct sockaddr *)&addr, NULL) != 0) {
-        return;
-    }
-
-    struct sockaddr_un svaddr, claddr;
-    int sfd = -1;
-    size_t msgLen;
-    char resp[SOCKET_BUF_SIZE];
-
-    if (verbose) {
-        LOGI("update traffic stat: tx: %" PRIu64 " rx: %" PRIu64 "", tx, rx);
-    }
-
-    snprintf(resp, SOCKET_BUF_SIZE, "stat: {\"%s\":%" PRIu64 "}", sockaddr_readable("%p", &addr), tx + rx);
-    msgLen = strlen(resp) + 1;
-
-    ss_addr_t ip_addr = { .host = NULL, .port = NULL };
-    parse_addr(manager_addr, &ip_addr);
-
-    if (ip_addr.host == NULL || ip_addr.port == NULL) {
-        sfd = socket(AF_UNIX, SOCK_DGRAM, 0);
-        if (sfd == -1) {
-            ERROR("stat_socket");
-            return;
-        }
-
-        memset(&claddr, 0, sizeof(struct sockaddr_un));
-        claddr.sun_family = AF_UNIX;
-        snprintf(claddr.sun_path, sizeof(claddr.sun_path), "/tmp/shadowsocks.%s", sockaddr_readable("%p", &addr));
-
-        unlink(claddr.sun_path);
-
-        if (bind(sfd, (struct sockaddr *)&claddr, sizeof(struct sockaddr_un)) == -1) {
-            ERROR("stat_bind");
-            close(sfd);
-            return;
-        }
-
-        memset(&svaddr, 0, sizeof(struct sockaddr_un));
-        svaddr.sun_family = AF_UNIX;
-        strncpy(svaddr.sun_path, manager_addr, sizeof(svaddr.sun_path) - 1);
-
-        if (sendto(sfd, resp, strlen(resp) + 1, 0, (struct sockaddr *)&svaddr,
-                   sizeof(struct sockaddr_un)) != msgLen) {
-            ERROR("stat_sendto");
-            close(sfd);
-            return;
-        }
-
-        unlink(claddr.sun_path);
-    } else {
-        struct sockaddr_storage storage;
-        memset(&storage, 0, sizeof(struct sockaddr_storage));
-        if (get_sockaddr(ip_addr.host, ip_addr.port, &storage, 1, ipv6first) == -1) {
-            ERROR("failed to parse the manager addr");
-            return;
-        }
-
-        sfd = socket(storage.ss_family, SOCK_DGRAM, 0);
-
-        if (sfd == -1) {
-            ERROR("stat_socket");
-            return;
-        }
-
-        size_t addr_len = get_sockaddr_len((struct sockaddr *)&storage);
-        if (sendto(sfd, resp, strlen(resp) + 1, 0, (struct sockaddr *)&storage,
-                   addr_len) != msgLen) {
-            ERROR("stat_sendto");
-            close(sfd);
-            return;
-        }
-    }
-
-    close(sfd);
-}
-
-#endif
-
-static void
-free_listeners(struct ev_loop *loop)
-{
-    listen_ctx_t *listener = NULL;
-    struct cork_dllist_item *curr, *next;
-    cork_dllist_foreach(&listeners, curr, next,
-                        listen_ctx_t, listener, entries) {
-        if (listener != NULL) {
-#ifndef __MINGW32__
-            if (manager_addr != NULL)
-                ev_timer_stop(loop, &stat_watcher);
-#endif
-            ev_io_stop(loop, &listener->io);
-            close(listener->fd);
-        }
-    }
-}
 
 static void
 report_addr(EV_P_ server_t *server, const char *info)
@@ -905,32 +781,7 @@ remote_send_cb(EV_P_ ev_io *w, int revents)
     }
 }
 
-static void
-signal_cb(EV_P_ ev_signal *w, int revents)
-{
-    if (revents & EV_SIGNAL) {
-        switch (w->signum) {
-#ifndef __MINGW32__
-        case SIGCHLD:
-            if (!is_plugin_running()) {
-                LOGE("plugin service exit unexpectedly");
-                ret_val = -1;
-            } else
-                return;
-#endif
-        case SIGINT:
-        case SIGTERM:
-            ev_signal_stop(EV_DEFAULT, &sigint_watcher);
-            ev_signal_stop(EV_DEFAULT, &sigterm_watcher);
-#ifndef __MINGW32__
-            ev_signal_stop(EV_DEFAULT, &sigchld_watcher);
-#endif
-            ev_unloop(EV_A_ EVUNLOOP_ALL);
-        }
-    }
-}
-
-static void
+void
 accept_cb(EV_P_ ev_io *w, int revents)
 {
     listen_ctx_t *listener = (listen_ctx_t *)w;
@@ -978,227 +829,18 @@ main(int argc, char **argv)
     USE_TTY();
     srand(time(NULL));
 
-    int i;
-    int plugin_enabled = 0;
     int pid_flags = 0;
     jconf_t conf  = jconf_default;
 
-    if (parse_argopts(&conf, argc, argv) != 0 ||
-        !(conf.remotes != NULL && conf.remote_num > 0))
-    {
+    if (parse_argopts(&conf, argc, argv) != 0) {
         usage();
         exit(EXIT_FAILURE);
-    }
-
-    if (conf.log) {
-        USE_LOGFILE(conf.log);
-        LOGI("enabled %slogging %s", conf.verbose ? "verbose " : "", conf.log);
-    }
-
-    if (conf.mtu > 0) {
-        LOGI("setting MTU to %d", conf.mtu);
-    }
-
-    if (conf.mptcp) {
-        LOGI("enabled multipath TCP");
-    }
-
-    if (conf.no_delay) {
-        LOGI("enabled TCP no-delay");
-    }
-
-    if (conf.mode != TCP_ONLY) {
-        LOGI("UDP relay enabled");
-    }
-
-    if (conf.mode == UDP_ONLY) {
-        LOGI("TCP relay disabled");
-    }
-
-    if (conf.ipv6_first) {
-        LOGI("prioritized IPv6 addresses in domain resolution");
-    }
-
-    if (conf.acl != NULL) {
-        LOGI("initializing acl...");
-        acl = !init_acl(&conf);
-    }
-
-#ifdef HAVE_SETRLIMIT
-    /*
-     * No need to check the return value here.
-     * We will show users an error message if setrlimit(2) fails.
-     */
-    if (conf.nofile > 1024) {
-        if (conf.verbose) {
-            LOGI("setting NOFILE to %d", conf.nofile);
-        }
-        set_nofile(conf.nofile);
-    }
-#endif
-
-    if (conf.fast_open) {
-#ifdef TCP_FASTOPEN
-        LOGI("using tcp fast open");
-#else
-        LOGE("tcp fast open is not supported by this environment");
-        conf.fast_open = 0;
-#endif
     }
 
     pid_flags = conf.pid_path != NULL;
     USE_SYSLOG(argv[0], pid_flags);
     if (pid_flags) {
         daemonize(conf.pid_path);
-    }
-
-    no_delay   = conf.no_delay;
-    fast_open  = conf.fast_open;
-    verbose    = conf.verbose;
-    ipv6first  = conf.ipv6_first;
-    manager_addr = conf.manager_addr;
-
-#ifdef __MINGW32__
-    winsock_init();
-#endif
-
-#ifndef __MINGW32__
-    // ignore SIGPIPE
-    signal(SIGPIPE, SIG_IGN);
-    signal(SIGABRT, SIG_IGN);
-#endif
-
-    ev_signal_init(&sigint_watcher, signal_cb, SIGINT);
-    ev_signal_init(&sigterm_watcher, signal_cb, SIGTERM);
-    ev_signal_start(EV_DEFAULT, &sigint_watcher);
-    ev_signal_start(EV_DEFAULT, &sigterm_watcher);
-#ifndef __MINGW32__
-    ev_signal_init(&sigchld_watcher, signal_cb, SIGCHLD);
-    ev_signal_start(EV_DEFAULT, &sigchld_watcher);
-#endif
-
-    struct ev_loop *loop = EV_DEFAULT;
-    resolv_init(loop, conf.nameserver, conf.ipv6_first);
-    if (conf.nameserver != NULL)
-        LOGI("using nameserver: %s", conf.nameserver);
-    port_service_init();
-
-    cork_dllist_init(&listeners);
-
-    for (i = 0; i < conf.remote_num; i++) {
-        ss_remote_t *r = conf.remotes[i];
-        char *host     = r->addr,
-             *port     = elvis(r->port, conf.remote_port),
-             *password = elvis(r->password, conf.password),
-             *key      = elvis(r->key, conf.key),
-             *method   = elvis(r->method, conf.method),
-             *iface    = elvis(r->iface, conf.iface),
-             *plugin   = elvis(r->plugin, conf.plugin),
-             *plugin_opts
-                       = elvis(r->plugin_opts, conf.plugin_opts);
-
-        if (port == NULL || method == NULL ||
-            (password == NULL && key == NULL))
-        {
-            usage();
-            exit(EXIT_FAILURE);
-        }
-
-        LOGI("[%d/%d] listening on %s:%s", i + 1, conf.remote_num, host, port);
-
-        LOGI("initializing ciphers... %s", method);
-        crypto_t *crypto = crypto_init(password, key, method);
-        if (crypto == NULL)
-            FATAL("failed to initialize ciphers");
-
-        /**
-         * "If node(host) is NULL, then the network address
-         * will be set to the loopback interface address," meaning that
-         * by default the server will bind to `lo' instead of 0.0.0.0.
-         *
-         * On Linux, when net.ipv6.bindv6only = 0 (the default),
-         * getaddrinfo(NULL) with AI_PASSIVE returns 0.0.0.0 and :: (in this order).
-         * AI_PASSIVE was meant to return a list of addresses to listen on,
-         * but it is impossible to listen on 0.0.0.0 and :: at the same time,
-         * unless `bindv6only' is enabled.
-         */
-        struct sockaddr_storage *storage =
-                ss_calloc(1, sizeof(struct sockaddr_storage));
-        if (get_sockaddr(host, port, storage, 1, conf.ipv6_first) == -1) {
-            FATAL("failed to resolve %s", host);
-        }
-
-        if (plugin != NULL) {
-            if (!plugin_enabled) {
-                init_plugin(MODE_SERVER);
-                plugin_enabled = 1;
-            }
-
-            uint16_t plugin_port = get_local_port();
-            switch (storage->ss_family) {
-                case AF_INET: {
-                    *(struct sockaddr_in *)storage = (struct sockaddr_in) {
-                        .sin_addr =
-                            (struct in_addr) { htonl(INADDR_LOOPBACK) },
-                        .sin_port = plugin_port
-                    };
-                } break;
-                case AF_INET6: {
-                    *(struct sockaddr_in6 *)storage = (struct sockaddr_in6) {
-                        .sin6_addr = in6addr_loopback,
-                        .sin6_port = plugin_port
-                    };
-                } break;
-            }
-
-            if (plugin_port == 0)
-                FATAL("failed to find a free port");
-
-            LOGI("plugin \"%s\" enabled", plugin);
-
-            int err = start_plugin(plugin, plugin_opts,                 // user-defined plugin options
-                                   sockaddr_readable("%a", storage),
-                                   sockaddr_readable("%p", storage),    // plugin destination override
-                                   host, port);
-            if (err)
-                FATAL("failed to start the plugin");
-        }
-
-        listen_ctx_t listen_ctx = {
-            .iface   = iface,
-            .addr    = storage,
-            .crypto  = crypto,
-            .timeout = atoi(conf.timeout),
-            .loop    = loop
-        };
-
-#ifndef __MINGW32__
-        if (conf.manager_addr != NULL) {
-            ev_timer_init(&listen_ctx.stat_watcher, stat_update_cb,
-                          UPDATE_INTERVAL, UPDATE_INTERVAL);
-            ev_timer_start(EV_DEFAULT, &listen_ctx.stat_watcher);
-        }
-#endif
-
-        if (conf.mode != UDP_ONLY) {
-            int listenfd = bind_and_listen(storage, IPPROTO_TCP, &listen_ctx);
-            if (listenfd != -1) {
-                if (fast_open)
-                    set_fastopen_passive(listenfd);
-                ev_io_init(&listen_ctx.io, accept_cb, listen_ctx.fd, EV_READ);
-                ev_io_start(loop, &listen_ctx.io);
-                cork_dllist_add(&listeners, &listen_ctx.entries);
-            }
-        }
-
-        // Setup UDP
-        if (conf.mode != TCP_ONLY) {
-            listen_ctx_t listen_ctx_dgram = listen_ctx;
-            int listenfd = bind_and_listen(storage, IPPROTO_UDP, &listen_ctx_dgram);
-            if ((listen_ctx_dgram.fd = listenfd) != -1) {
-                init_udprelay(&listen_ctx_dgram);
-            }
-        }
     }
 
 #ifndef __MINGW32__
@@ -1212,40 +854,9 @@ main(int argc, char **argv)
     }
 #endif
 
-    // Init connections
-    cork_dllist_init(&connections);
-
-    // start ev loop
-    ev_run(loop, 0);
-
-    if (verbose) {
-        LOGI("closed gracefully");
-    }
-
-#ifndef __MINGW32__
-    if (conf.manager_addr != NULL) {
-        ev_timer_stop(loop, &stat_watcher);
-    }
-#endif
-
-    if (plugin_enabled) {
-        stop_plugin();
-    }
-
-    resolv_shutdown(loop);
-
-    if (conf.mode != UDP_ONLY) {
-        free_listeners(loop);
-        free_connections(loop);
-    }
-
-    if (conf.mode != TCP_ONLY) {
-        free_udprelay(loop);
-    }
-
-#ifdef __MINGW32__
-    winsock_cleanup();
-#endif
-
-    return ret_val;
+    no_delay  = conf.no_delay;
+    fast_open = conf.fast_open;
+    verbose   = conf.verbose;
+    ipv6first = conf.ipv6_first;
+    return start_relay(&conf, NULL, NULL);
 }
