--- src/server.h
@@ -1,115 +1,0 @@
-/*
- * server.h - Define shadowsocks server's buffers and callbacks
- *
- * Copyright (C) 2013 - 2019, Max Lv <max.c.lv@gmail.com>
- *
- * This file is part of the shadowsocks-libev.
- *
- * shadowsocks-libev is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3 of the License, or
- * (at your option) any later version.
- *
- * shadowsocks-libev is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with shadowsocks-libev; see the file COPYING. If not, see
- * <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _SERVER_H
-#define _SERVER_H
-
-#include <time.h>
-#include <libcork/ds.h>
-
-#ifdef HAVE_LIBEV_EV_H
-#include <libev/ev.h>
-#else
-#include <ev.h>
-#endif
-
-#ifdef __MINGW32__
-#include "winsock.h"
-#endif
-
-#include "crypto.h"
-#include "jconf.h"
-#include "resolv.h"
-#include "netutils.h"
-
-#include "common.h"
-
-typedef struct server_ctx {
-    ev_io io;
-    ev_timer watcher;
-    int connected;
-    struct server *server;
-} server_ctx_t;
-
-#ifdef USE_NFCONNTRACK_TOS
-
-#include <libnetfilter_conntrack/libnetfilter_conntrack.h>
-#include <libnetfilter_conntrack/libnetfilter_conntrack_tcp.h>
-
-struct dscptracker {
-    struct nf_conntrack *ct;
-    long unsigned int mark;
-    unsigned int dscp;
-    unsigned int packet_count;
-};
-
-#endif
-
-struct query;
-
-typedef struct server {
-    int fd;
-    int stage;
-    int frag;
-
-    buffer_t *buf;
-
-    crypto_t *crypto;
-    cipher_ctx_t *e_ctx;
-    cipher_ctx_t *d_ctx;
-    struct server_ctx *recv_ctx;
-    struct server_ctx *send_ctx;
-    struct listen_ctx *listen_ctx;
-    struct remote *remote;
-
-    struct query *query;
-
-    struct cork_dllist_item entries;
-#ifdef USE_NFCONNTRACK_TOS
-    struct dscptracker *tracker;
-#endif
-} server_t;
-
-typedef struct query {
-    server_t *server;
-    char hostname[MAX_HOSTNAME_LEN];
-} query_t;
-
-typedef struct remote_ctx {
-    ev_io io;
-    int connected;
-    struct remote *remote;
-} remote_ctx_t;
-
-typedef struct remote {
-    int fd;
-#ifdef TCP_FASTOPEN_WINSOCK
-    OVERLAPPED olap;
-    int connect_ex_done;
-#endif
-    buffer_t *buf;
-    struct remote_ctx *recv_ctx;
-    struct remote_ctx *send_ctx;
-    struct server *server;
-} remote_t;
-
-#endif // _SERVER_H
