--- src/tunnel.h
@@ -1,79 +1,0 @@
-/*
- * tunnel.h - Define tunnel's buffers and callbacks
- *
- * Copyright (C) 2013 - 2019, Max Lv <max.c.lv@gmail.com>
- *
- * This file is part of the shadowsocks-libev.
- *
- * shadowsocks-libev is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3 of the License, or
- * (at your option) any later version.
- *
- * shadowsocks-libev is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with shadowsocks-libev; see the file COPYING. If not, see
- * <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _TUNNEL_H
-#define _TUNNEL_H
-
-#ifdef HAVE_LIBEV_EV_H
-#include <libev/ev.h>
-#else
-#include <ev.h>
-#endif
-
-#ifdef __MINGW32__
-#include "winsock.h"
-#endif
-
-#include "crypto.h"
-#include "jconf.h"
-
-#include "common.h"
-
-typedef struct server_ctx {
-    ev_io io;
-    int connected;
-    struct server *server;
-} server_ctx_t;
-
-typedef struct server {
-    int fd;
-
-    buffer_t *buf;
-
-    struct server_ctx *recv_ctx;
-    struct server_ctx *send_ctx;
-    struct remote *remote;
-
-    struct listen_ctx *listener;
-} server_t;
-
-typedef struct remote_ctx {
-    ev_io io;
-    ev_timer watcher;
-    int connected;
-    struct remote *remote;
-} remote_ctx_t;
-
-typedef struct remote {
-    int fd;
-#ifdef TCP_FASTOPEN_WINSOCK
-    OVERLAPPED olap;
-    int connect_ex_done;
-#endif
-    buffer_t *buf;
-    struct remote_ctx *recv_ctx;
-    struct remote_ctx *send_ctx;
-    struct server *server;
-    struct sockaddr *addr;
-} remote_t;
-
-#endif // _TUNNEL_H
